requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-migrate==4.0.7
authlib==1.3.1
bcrypt==4.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
gunicorn==22.0.0
stripe==9.9.0
wtforms==3.1.2
email-validator==2.2.0

anthropic==0.40.0
